- created extension function to wrap Movie - separated the models into folders
