+++ app/src/main/java/com/example/topmovies/database/MovieDatabase.kt
@@ -4,8 +4,8 @@ import androidx.room.Database
 import androidx.room.RoomDatabase
 import com.example.topmovies.database.dao.MovieDetailsDao
 import com.example.topmovies.database.dao.MoviesDao
-import com.example.topmovies.models.MovieDetailsEntity
-import com.example.topmovies.models.MovieEntity
+import com.example.topmovies.models.entity.MovieDetailsEntity
+import com.example.topmovies.models.entity.MovieEntity
 
 @Database(entities = [MovieEntity::class, MovieDetailsEntity::class], version = 1)
 abstract class MovieDatabase : RoomDatabase() {
