+++ app/src/main/java/com/example/topmovies/database/dao/MovieDetailsDao.kt
@@ -2,7 +2,7 @@ package com.example.topmovies.database.dao
 
 import androidx.room.Dao
 import androidx.room.Query
-import com.example.topmovies.models.MovieDetailsEntity
+import com.example.topmovies.models.entity.MovieDetailsEntity
 
 @Dao
 abstract class MovieDetailsDao : BaseDao<MovieDetailsEntity> {
