+++ app/src/main/java/com/example/topmovies/domain/GetMovieDetailsUseCase.kt
@@ -1,7 +1,9 @@
 package com.example.topmovies.domain
 
 import android.os.Handler
-import com.example.topmovies.models.MovieDetails
+import com.example.topmovies.models.domain.MovieDetails
+import com.example.topmovies.models.domain.toMoviesDetailsEntity
+import com.example.topmovies.models.entity.toMovieDetails
 import com.example.topmovies.repository.MovieRepository
 import java.util.concurrent.ExecutorService
 
@@ -18,7 +20,9 @@ class GetMovieDetailsUseCase(
             if (resultApi.error.isNotEmpty()) {
                 resultApi
             } else {
-                repository.insertMovieDetailsEntity(resultApi.value.toMovieDetailsEntity())
+                repository.insertMovieDetailsEntity(
+                    resultApi.value.toMoviesDetailsEntity()
+                )
                 Result(resultApi.value)
             }
         } else {
