+++ app/src/main/java/com/example/topmovies/fragment/MoviesFragment.kt
@@ -10,7 +10,7 @@ import androidx.navigation.fragment.findNavController
 import com.example.topmovies.R
 import com.example.topmovies.adapter.MoviesAdapter
 import com.example.topmovies.databinding.FragmentMoviesBinding
-import com.example.topmovies.models.Movie
+import com.example.topmovies.models.domain.Movie
 import com.example.topmovies.unit.EnumScreen
 import com.example.topmovies.viewmodel.MovieViewModel
 import org.koin.androidx.viewmodel.ext.android.sharedViewModel
