+++ app/src/main/java/com/example/topmovies/models/response/MovieObjectResponse.kt
@@ -1,0 +1,5 @@
+package com.example.topmovies.models.response
+
+data class MovieObjectResponse(
+    val items: List<MovieResponse>, val errorMessage: String
+)
