+++ app/src/main/java/com/example/topmovies/retrofit/MoviesApi.kt
@@ -1,7 +1,7 @@
 package com.example.topmovies.retrofit
 
-import com.example.topmovies.models.MovieDetailsApi
-import com.example.topmovies.models.MovieObject
+import com.example.topmovies.models.response.MovieDetailsResponse
+import com.example.topmovies.models.response.MovieObjectResponse
 import com.example.topmovies.unit.MOVIE_DETAILED
 import com.example.topmovies.unit.TOP_100_MOVIES
 import retrofit2.Call
@@ -11,17 +11,17 @@ import retrofit2.http.Path
 interface MoviesApi {
     
     @GET("$TOP_100_MOVIES/{apiKey}")
-    fun getMovies(@Path("apiKey") apiKey: String): Call<MovieObject>
+    fun getMovies(@Path("apiKey") apiKey: String): Call<MovieObjectResponse>
     
     /** This method is for test, because 100 responses are not enough **/
     @GET("https://86e6737b-9ab8-444a-9c53-bdfa86309d8f.mock.pstmn.io/movies")
-    fun getMovies(): Call<MovieObject>
+    fun getMovies(): Call<MovieObjectResponse>
     
     @GET("$MOVIE_DETAILED/{apiKey}/{movieId}")
     fun getMovieDetails(@Path("apiKey") apiKey: String, @Path("movieId") movieId: String):
-            Call<MovieDetailsApi>
+            Call<MovieDetailsResponse>
     
     /** This method is for test, because 100 responses are not enough **/
     @GET("https://86e6737b-9ab8-444a-9c53-bdfa86309d8f.mock.pstmn.io/details")
-    fun getMovieDetails(): Call<MovieDetailsApi>
+    fun getMovieDetails(): Call<MovieDetailsResponse>
 }
