+++ app/src/main/java/com/example/topmovies/viewmodel/MovieViewModel.kt
@@ -7,7 +7,7 @@ import com.example.topmovies.domain.LoadMoviesUseCase
 import com.example.topmovies.domain.UpdateFavoriteMovieUseCase
 import com.example.topmovies.domain.UpdateFavoriteMovieUseCase.Params
 import com.example.topmovies.domain.UseCase.None
-import com.example.topmovies.models.Movie
+import com.example.topmovies.models.domain.Movie
 import com.example.topmovies.unit.EnumScreen
 
 class MovieViewModel constructor(
